+++ crates/networking/p2p/peer_handler.rs
@@ -35,8 +35,8 @@ use crate::{
 };
 use tracing::{debug, info, warn};
 pub const PEER_REPLY_TIMEOUT: Duration = Duration::from_secs(15);
-pub const PEER_SELECT_RETRY_ATTEMPTS: usize = 3;
-pub const REQUEST_RETRY_ATTEMPTS: usize = 5;
+pub const PEER_SELECT_RETRY_ATTEMPTS: u32 = 3;
+pub const REQUEST_RETRY_ATTEMPTS: u32 = 5;
 pub const MAX_RESPONSE_BYTES: u64 = 512 * 1024;
 pub const HASH_MAX: H256 = H256([0xFF; 32]);
